requirements.txt
#Name                    Version       
cuda100                   1.0           
cudatoolkit               10.1.243      
face-alignment            1.0.0         
face-recognition-models   0.3.0         
facenet-pytorch           2.2.9         
imagecodecs               2020.2.18     
imageio                   2.8.0         
libopencv                 3.4.2         
matplotlib                3.1.2         
matplotlib-base           3.1.2         
mkl                       2019.4        
mkl-service               2.3.0         
numpy                     1.18.1        
numpy-base                1.18.1        
opencv                    3.4.2         
opencv-python             4.1.0.25      
python                    3.7.6         
pytorch                   1.4.0         
pytorch-wavelets          1.2.3         
scikit-image              0.17.2        
scikit-learn              0.23.2        
scipy                     1.3.2         
tensorboardx              2.1           
torchvision               0.3.0         
tqdm                      4.42.0        

